removed some strcpy
git-svn-id: svn://svn.savannah.gnu.org/myserver/trunk@1164 c4c922ea-a70d-0410-937d-ec1daf91936c

+++ myserverweb/source/securestr.cpp
@@ -23,7 +23,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 /*!
  *Secure string concatenate routine.
  */
-unsigned int myserver_strlcat(char *destination, const char *source, size_t size)
+unsigned int myserver_strlcat(char *destination, const char *source, unsigned int size)
 {
 	char *dstptr=destination;
 	size_t dstlen,tocopy=size;
